Rename CitationContextScraper file for consistency
